+++ Libs/libMacGitverCore/CMakeLists.txt
@@ -44,6 +44,7 @@ SET( SRC_FILES
 
     RepoMan/Data/Base.cpp
     RepoMan/Data/Reference.cpp
+    RepoMan/Data/Branch.cpp
     RepoMan/Data/Tag.cpp
 
     RepoMan/Private/Dumper.cpp
