+++ Libs/libMacGitverCore/RepoMan/Data/Branch.cpp
@@ -1,0 +1,95 @@
+/*
+ * MacGitver
+ * Copyright (C) 2012-2015 The MacGitver-Developers <dev@macgitver.org>
+ *
+ * (C) Sascha Cunz <sascha@cunz-rad.com>
+ * (C) Cunz RaD Ltd.
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License (Version 2) as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+ * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program; if
+ * not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include "RepoMan/Data/BranchData.hpp"
+#include "RepoMan/Events.hpp"
+
+#include "RepoMan/Private/Dumper.hpp"
+
+namespace RM
+{
+
+    namespace Internal
+    {
+
+        BranchPrivate::BranchPrivate(Branch* pub, const Git::Reference& ref)
+            : RefPrivate(pub, BranchType, ref)
+            , mHasUpstream(false)
+            , mAheadCount(0)
+            , mBehindCount(0)
+        {
+        }
+
+        ObjTypes BranchPrivate::objType() const
+        {
+            return ObjTypes::Branch;
+        }
+
+        void BranchPrivate::postCreation()
+        {
+            if (!repoEventsBlocked()) {
+                Events::self()->branchCreated(repository(), pub<Branch>());
+            }
+
+            RefPrivate::postCreation();
+        }
+
+        void BranchPrivate::preTerminate()
+        {
+            if (!repoEventsBlocked()) {
+                Events::self()->branchAboutToBeDeleted(repository(), pub<Branch>());
+            }
+
+            RefPrivate::preTerminate();
+        }
+
+        void BranchPrivate::dumpSelf(Internal::Dumper& dumper) const
+        {
+            dumper.addLine(QString(QStringLiteral("Branch 0x%1 - %2"))
+                           .arg(quintptr(mPub),0,16)
+                           .arg(mName));
+        }
+
+        QString BranchPrivate::objectTypeName() const
+        {
+            return QStringLiteral("Branch");
+        }
+
+        void BranchPrivate::emitMoved()
+        {
+            if (!repoEventsBlocked()) {
+                Events::self()->refMoved(mRepo, pub<Ref>());
+                Events::self()->branchMoved(mRepo, pub<Branch>());
+            }
+        }
+
+        bool BranchPrivate::refreshDetails(const Git::Reference& ref)
+        {
+
+            return RefPrivate::refreshDetails(ref);
+        }
+
+        bool BranchPrivate::inherits(ObjTypes type) const
+        {
+            return type == ObjTypes::Branch || RefPrivate::inherits(type);
+        }
+
+    }
+
+}
