+++ Libs/libMacGitverCore/RepoMan/Frontend/Branch.cpp
@@ -74,68 +74,4 @@ namespace RM
         return h && h->is(this);
     }
 
-    //-- BranchPrivate -----------------------------------------------------------------------------
-
-    BranchPrivate::BranchPrivate(Branch* pub, const Git::Reference& ref)
-        : RefPrivate(pub, BranchType, ref)
-        , mHasUpstream(false)
-        , mAheadCount(0)
-        , mBehindCount(0)
-    {
-    }
-
-    ObjTypes BranchPrivate::objType() const
-    {
-        return ObjTypes::Branch;
-    }
-
-    void BranchPrivate::postCreation()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->branchCreated(repository(), pub<Branch>());
-        }
-
-        RefPrivate::postCreation();
-    }
-
-    void BranchPrivate::preTerminate()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->branchAboutToBeDeleted(repository(), pub<Branch>());
-        }
-
-        RefPrivate::preTerminate();
-    }
-
-    void BranchPrivate::dumpSelf(Internal::Dumper& dumper) const
-    {
-        dumper.addLine(QString(QStringLiteral("Branch 0x%1 - %2"))
-                       .arg(quintptr(mPub),0,16)
-                       .arg(mName));
-    }
-
-    QString BranchPrivate::objectTypeName() const
-    {
-        return QStringLiteral("Branch");
-    }
-
-    void BranchPrivate::emitMoved()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->refMoved(mRepo, pub<Ref>());
-            Events::self()->branchMoved(mRepo, pub<Branch>());
-        }
-    }
-
-    bool BranchPrivate::refreshDetails(const Git::Reference& ref)
-    {
-
-        return RefPrivate::refreshDetails(ref);
-    }
-
-    bool BranchPrivate::inherits(ObjTypes type) const
-    {
-        return type == ObjTypes::Branch || RefPrivate::inherits(type);
-    }
-
 }
